requirements.txt
Flask==0.12.2
gunicorn==19.7.1
requests==2.18.3
simplejson==3.11.1
python-telegram-bot==7.0.1